lte/alt1250: Change usrsock xid from uint64_t to uint32_t
Fix the type to match usrsock.h has been changed.

+++ lte/alt1250/alt1250_devevent.c
@@ -53,7 +53,7 @@
 
 static int handle_replypkt(FAR struct alt1250_s *dev,
   FAR struct alt_container_s *reply,
-  FAR int32_t *usock_result, uint64_t *usock_xid,
+  FAR int32_t *usock_result, uint32_t *usock_xid,
   FAR struct usock_ackinfo_s *ackinfo)
 {
   int ret;
@@ -86,7 +86,7 @@ static int perform_alt1250_reply(FAR struct alt1250_s *dev,
 {
   int ret = REP_NO_ACK;
   int32_t ack_result = OK;
-  uint64_t ack_xid = 0;
+  uint32_t ack_xid = 0;
   struct usock_ackinfo_s ackinfo;
 
   ret = handle_replypkt(dev, container, &ack_result, &ack_xid, &ackinfo);
@@ -131,7 +131,7 @@ static int handle_intentional_reset(FAR struct alt1250_s *dev)
 {
   if (dev->lwm2m_apply_xid >= 0)
     {
-      usockif_sendack(dev->usockfd, 0, (uint64_t)dev->lwm2m_apply_xid, false);
+      usockif_sendack(dev->usockfd, 0, (uint32_t)dev->lwm2m_apply_xid, false);
       dev->lwm2m_apply_xid = -1;
     }
