+++ lte/alt1250/usock_handlers/alt1250_sms.c
@@ -102,7 +102,7 @@ static int postproc_smsinit(FAR struct alt1250_s *dev,
                             FAR struct alt_container_s *reply,
                             FAR struct usock_s *usock,
                             FAR int32_t *usock_result,
-                            FAR uint64_t *usock_xid,
+                            FAR uint32_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo,
                             unsigned long arg);
 
@@ -110,7 +110,7 @@ static int postproc_smsinit_reopen(FAR struct alt1250_s *dev,
                                    FAR struct alt_container_s *reply,
                                    FAR struct usock_s *usock,
                                    FAR int32_t *usock_result,
-                                   FAR uint64_t *usock_xid,
+                                   FAR uint32_t *usock_xid,
                                    FAR struct usock_ackinfo_s *ackinfo,
                                    unsigned long arg);
 
@@ -118,7 +118,7 @@ static int postproc_smsfin(FAR struct alt1250_s *dev,
                            FAR struct alt_container_s *reply,
                            FAR struct usock_s *usock,
                            FAR int32_t *usock_result,
-                           FAR uint64_t *usock_xid,
+                           FAR uint32_t *usock_xid,
                            FAR struct usock_ackinfo_s *ackinfo,
                            unsigned long arg);
 
@@ -126,7 +126,7 @@ static int postproc_smsfin_reopen(FAR struct alt1250_s *dev,
                                   FAR struct alt_container_s *reply,
                                   FAR struct usock_s *usock,
                                   FAR int32_t *usock_result,
-                                  FAR uint64_t *usock_xid,
+                                  FAR uint32_t *usock_xid,
                                   FAR struct usock_ackinfo_s *ackinfo,
                                   unsigned long arg);
 
@@ -134,7 +134,7 @@ static int postproc_smssend(FAR struct alt1250_s *dev,
                             FAR struct alt_container_s *reply,
                             FAR struct usock_s *usock,
                             FAR int32_t *usock_result,
-                            FAR uint64_t *usock_xid,
+                            FAR uint32_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo,
                             unsigned long arg);
 
@@ -142,7 +142,7 @@ static int postproc_smsdelete(FAR struct alt1250_s *dev,
                               FAR struct alt_container_s *reply,
                               FAR struct usock_s *usock,
                               FAR int32_t *usock_result,
-                              FAR uint64_t *usock_xid,
+                              FAR uint32_t *usock_xid,
                               FAR struct usock_ackinfo_s *ackinfo,
                               unsigned long arg);
 
@@ -466,7 +466,7 @@ static int postproc_smsinit(FAR struct alt1250_s *dev,
                             FAR struct alt_container_s *reply,
                             FAR struct usock_s *usock,
                             FAR int32_t *usock_result,
-                            FAR uint64_t *usock_xid,
+                            FAR uint32_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo,
                             unsigned long arg)
 {
@@ -497,7 +497,7 @@ static int postproc_smsinit_reopen(FAR struct alt1250_s *dev,
                                    FAR struct alt_container_s *reply,
                                    FAR struct usock_s *usock,
                                    FAR int32_t *usock_result,
-                                   FAR uint64_t *usock_xid,
+                                   FAR uint32_t *usock_xid,
                                    FAR struct usock_ackinfo_s *ackinfo,
                                    unsigned long arg)
 {
@@ -524,7 +524,7 @@ static int postproc_smsfin(FAR struct alt1250_s *dev,
                            FAR struct alt_container_s *reply,
                            FAR struct usock_s *usock,
                            FAR int32_t *usock_result,
-                           FAR uint64_t *usock_xid,
+                           FAR uint32_t *usock_xid,
                            FAR struct usock_ackinfo_s *ackinfo,
                            unsigned long arg)
 {
@@ -546,7 +546,7 @@ static int postproc_smsfin_reopen(FAR struct alt1250_s *dev,
                                   FAR struct alt_container_s *reply,
                                   FAR struct usock_s *usock,
                                   FAR int32_t *usock_result,
-                                  FAR uint64_t *usock_xid,
+                                  FAR uint32_t *usock_xid,
                                   FAR struct usock_ackinfo_s *ackinfo,
                                   unsigned long arg)
 {
@@ -577,7 +577,7 @@ static int postproc_smssend(FAR struct alt1250_s *dev,
                             FAR struct alt_container_s *reply,
                             FAR struct usock_s *usock,
                             FAR int32_t *usock_result,
-                            FAR uint64_t *usock_xid,
+                            FAR uint32_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo,
                             unsigned long arg)
 {
@@ -611,7 +611,7 @@ static int postproc_smsdelete(FAR struct alt1250_s *dev,
                               FAR struct alt_container_s *reply,
                               FAR struct usock_s *usock,
                               FAR int32_t *usock_result,
-                              FAR uint64_t *usock_xid,
+                              FAR uint32_t *usock_xid,
                               FAR struct usock_ackinfo_s *ackinfo,
                               unsigned long arg)
 {
@@ -950,7 +950,7 @@ int alt1250_sms_recv(FAR struct alt1250_s *dev,
 int usockreq_ioctl_sms(FAR struct alt1250_s *dev,
                        FAR struct usrsock_request_buff_s *req,
                        FAR int32_t *usock_result,
-                       FAR uint64_t *usock_xid,
+                       FAR uint32_t *usock_xid,
                        FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_ioctl_s *request = &req->request.ioctl_req;
